Update notices; Standardize preamble...
git-svn-id: file:///home/jj/hercules.svn/trunk@5961 956126f8-22a0-4046-8f4a-272fa8102e63

+++ config.c
@@ -1,5 +1,9 @@
 /* CONFIG.C     (c) Copyright Jan Jaeger, 2000-2010                  */
 /*              Device configuration functions                       */
+/*                                                                   */
+/*   Released under "The Q Public License Version 1"                 */
+/*   (http://www.hercules-390.org/herclic.html) as modifications to  */
+/*   Hercules.                                                       */
 
 // $Id$
 
@@ -414,7 +418,7 @@ int     i;                              /* Loop index                */
         dev->argv = NULL;
 
     /* Call the device handler initialization function */
-    rc = (dev->hnd->init)(dev, addargc, addargv);
+    rc = (int)(dev->hnd->init)(dev, addargc, addargv);
 
     if (rc < 0)
     {
